Fix I2C device property type mismatch
The dump inventory tool was failing to collect I2C device inventory
properties and logging "Invalid data type received" while reading
Address and Bus properties.

The I2CDevice interface defines these properties with the D-Bus size
type, which is represented as a 64-bit unsigned value on the target
platform. Update the I2C device property handling from uint32_t to
uint64_t to match the existing DbusVariantType and correctly read the
inventory properties during dump inventory operation.

```
Test:
Applied patch on pST simics.
Observed I2C device properties are dumped using dump inventory option.
```

+++ vpd-tool/src/vpd_tool.cpp
@@ -279,7 +279,7 @@ void VpdTool::populateFruJson(
         if (l_interface == constants::i2cDeviceInf)
         {
             const std::vector<std::string> l_properties = {"Address", "Bus"};
-            populateInterfaceJson<uint32_t>(i_inventoryObjPath,
+            populateInterfaceJson<uint64_t>(i_inventoryObjPath,
                                             constants::i2cDeviceInf,
                                             l_properties, io_fruJsonObject);
             continue;
